Improve chapter title and arc presentation
This introduces arc title pages, and a hierarchical table of contents which nests the chapters under the arcs, powered by the new chapter-data/ files.

The chapter titles are configurable, using the new --chapter-titles option explained in the README.

Closes #54.

As a side-effect, this introduces a style sheet included by all chapters and arc title pages. This will help with #55.